Replace regex IP validation with ipaddress stdlib in EnrichmentSerializer. Closes #881
- Create greedybear/utils.py with shared validation helpers (is_ip_address, is_sha256hash)
- Replace REGEX_IP with is_ip_address() for proper IPv4/IPv6 validation
- Add .strip() with write-back for whitespace handling
- Improve domain validation to require at least one alphabetic character
- Update all import sites to use greedybear.utils
- Remove moved functions and unused imports from api/views/utils.py
- Add regression tests for invalid IPs, valid IPv6, whitespace, and domains
