+++ api/views/cowrie_session.py
@@ -16,9 +16,9 @@ from rest_framework.decorators import (
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
-from api.views.utils import is_ip_address, is_sha256hash
 from greedybear.consts import GET
 from greedybear.models import CommandSequence, CowrieSession, Statistics, ViewType
+from greedybear.utils import is_ip_address, is_sha256hash
 
 logger = logging.getLogger(__name__)
