+++ api/views/utils.py
@@ -2,9 +2,7 @@
 # See the file 'LICENSE' for copying permission.
 import csv
 import logging
-import re
 from datetime import datetime, timedelta
-from ipaddress import ip_address
 
 import feedparser
 import requests
@@ -307,43 +305,6 @@ def feeds_response(iocs, feed_params, valid_feed_types, dict_only=False, verbose
                 return Response(resp_data, status=status.HTTP_200_OK)
         case _:
             return HttpResponseBadRequest()
-
-
-def is_ip_address(string: str) -> bool:
-    """
-    Validate if a string is a valid IP address (IPv4 or IPv6).
-
-    Uses the ipaddress module to perform validation. This function properly
-    handles both IPv4 addresses and IPv6 addresses.
-
-    Args:
-        string: The string to validate as an IP address
-
-    Returns:
-        bool: True if the string is a valid IP address, False otherwise
-    """
-    try:
-        ip_address(string)
-    except ValueError:
-        return False
-    return True
-
-
-def is_sha256hash(string: str) -> bool:
-    """
-    Validate if a string is a valid SHA-256 hash.
-
-    A SHA-256 hash is a string of exactly 64 hexadecimal characters
-    (0-9, a-f, A-F). This function checks if the input string matches
-    this pattern using a regular expression.
-
-    Args:
-        string: The string to validate as a SHA-256 hash
-
-    Returns:
-        bool: True if the string is a valid SHA-256 hash, False otherwise
-    """
-    return bool(re.fullmatch(r"^[A-Fa-f0-9]{64}$", string))
 
 
 def asn_aggregated_queryset(iocs_qs, request, feed_params):
